Proof of Concept: creating Spreadsheet from Cells detected using fancy algos
1. There's some weirdness in ruling.rb around dealing with not-quite-intersecting lines.
2. Also, add_merged_cells needs to get called for those sorts of PDFs. Or maybe there's some
issue with intersection points that obviates the need for it (but loses some information
about merged cells). This may be connected to (1) -- e.g. extended lines "intersect" when they
"in fact" just abut each other (e.g. they really make a T shape, but because they've been
extended, they make a + shape). Not sure how to deal with this immediately, but results are
nevertheless quite good on my one test PDF.